requirements.txt
streamlit
langchain-openai
langchain-core
python-dotenv
langchain-community
networkx
pyvis